requirements.txt
asn1crypto==1.3.0
astroid==2.3.3
azure-common==1.1.24
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.11.12
botocore==1.14.12
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
docutils==0.15.2
idna==2.8
ijson==2.6.1
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oscrypto==1.2.0
pycparser==2.19
pycryptodomex==3.9.6
pygrametl==2.6
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
snowflake-connector-python==2.2.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
